Remove `table-logger` dependency (kubeflow/trainer#1544)
Closes https://github.com/kubeflow/training-operator/issues/1446.

This removes the dependency on `table-logger`. None of the various
alternatives mentioned in that issues were drop-in replacements, so
given the simplicity of the use case here the dependency is replaced
with a small class utilizing `str.format`. Note that the class was
written without referring to the `table-logger` source at all, which
would be a license violation.

I confirmed that you can now run the sample notebook

https://github.com/kubeflow/training-operator/blob/38ac5390872c43bd54a7e9deedee77d1558d45d4/sdk/python/examples/kubeflow-tfjob-sdk.ipynb

in modern IPython.